Extract OsFormField into its own component
